# kv: implement dist sender optimizations