--- a/pkg/kv/dist_sender.go
+++ b/pkg/kv/dist_sender.go
@@ -377,11 +377,6 @@ func (ds *DistSender) sendRPC(
 	// RangeNotFoundErrors.
 	ba.RangeID = rangeID
 
-	// A given RPC may generate retries to multiple replicas, but as soon as we
-	// get a response from one we want to cancel those other RPCs.
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
 	tracing.AnnotateTrace()
 	defer tracing.AnnotateTrace()
 
@@ -652,6 +647,28 @@ type response struct {
 func (ds *DistSender) divideAndSendBatchToRanges(
 	ctx context.Context, ba roachpb.BatchRequest, rs roachpb.RSpan, batchIdx int,
 ) (br *roachpb.BatchResponse, pErr *roachpb.Error) {
+	// Get initial seek key depending on direction of iteration.
+	var scanDir ScanDirection
+	var seekKey roachpb.RKey
+	if !ba.IsReverse() {
+		scanDir = Ascending
+		seekKey = rs.Key
+	} else {
+		scanDir = Descending
+		seekKey = rs.EndKey
+	}
+	ri := NewRangeIterator(ds)
+	ri.Seek(ctx, seekKey, scanDir)
+	if !ri.Valid() {
+		return nil, ri.Error()
+	}
+	// Take the fast path if this batch fits within a single range.
+	if !ri.NeedAnother(rs) {
+		ba.SetNewRequest()
+		resp := ds.sendPartialBatch(ctx, ba, rs, ri.Desc(), ri.Token(), batchIdx, false /* needsTruncate */)
+		return resp.reply, resp.pErr
+	}
+
 	// Make an empty slice of responses which will be populated with responses
 	// as they come in via Combine().
 	br = &roachpb.BatchResponse{
@@ -661,7 +678,6 @@ func (ds *DistSender) divideAndSendBatchToRanges(
 	// implicated in the span (rs) and combines them into a single
 	// BatchResponse when finished.
 	var responseChs []chan response
-	var seekKey roachpb.RKey
 	var couldHaveSkippedResponses bool
 	defer func() {
 		if r := recover(); r != nil {
@@ -702,18 +718,7 @@ func (ds *DistSender) divideAndSendBatchToRanges(
 		}
 	}()
 
-	// Get initial seek key depending on direction of iteration.
-	var scanDir ScanDirection
-	if !ba.IsReverse() {
-		scanDir = Ascending
-		seekKey = rs.Key
-	} else {
-		scanDir = Descending
-		seekKey = rs.EndKey
-	}
-	// Send the request to one range per iteration.
-	ri := NewRangeIterator(ds)
-	for ri.Seek(ctx, seekKey, scanDir); ri.Valid(); ri.Seek(ctx, seekKey, scanDir) {
+	for ; ri.Valid(); ri.Seek(ctx, seekKey, scanDir) {
 		// Increase the sequence counter only once before sending RPCs to
 		// the ranges involved in this chunk of the batch (as opposed to
 		// for each RPC individually). On RPC errors, there's no guarantee
@@ -801,7 +806,7 @@ func (ds *DistSender) divideAndSendBatchToRanges(
 		} else {
 			// Send synchronously if there is no parallel capacity left, there's a
 			// max results limit, or this is the final request in the span.
-			resp := ds.sendPartialBatch(ctx, ba, rs, ri.Desc(), ri.Token(), batchIdx)
+			resp := ds.sendPartialBatch(ctx, ba, rs, ri.Desc(), ri.Token(), batchIdx, true /* needsTruncate */)
 			responseCh <- resp
 			if resp.pErr != nil {
 				return
@@ -868,7 +873,7 @@ func (ds *DistSender) sendPartialBatchAsync(
 		ds.asyncSenderSem, false, /* !wait */
 		func(ctx context.Context) {
 			atomic.AddInt32(&ds.asyncSenderCount, 1)
-			responseCh <- ds.sendPartialBatch(ctx, ba, rs, desc, evictToken, batchIdx)
+			responseCh <- ds.sendPartialBatch(ctx, ba, rs, desc, evictToken, batchIdx, true /* needsTruncate */)
 		},
 	); err != nil {
 		return false
@@ -884,7 +889,9 @@ func (ds *DistSender) sendPartialBatchAsync(
 // replicas, we backoff and retry by refetching the range
 // descriptor. If the underlying range seems to have split, we
 // recursively invoke divideAndSendBatchToRanges to re-enumerate the
-// ranges in the span and resend to each.
+// ranges in the span and resend to each. If needsTruncate is true,
+// the supplied batch and span must be truncated to the supplied range
+// descriptor.
 func (ds *DistSender) sendPartialBatch(
 	ctx context.Context,
 	ba roachpb.BatchRequest,
@@ -892,6 +899,7 @@ func (ds *DistSender) sendPartialBatch(
 	desc *roachpb.RangeDescriptor,
 	evictToken *EvictionToken,
 	batchIdx int,
+	needsTruncate bool,
 ) response {
 	if batchIdx == 1 {
 		ds.metrics.PartialBatchCount.Inc(2) // account for first batch
@@ -900,23 +908,27 @@ func (ds *DistSender) sendPartialBatch(
 	}
 	var reply *roachpb.BatchResponse
 	var pErr *roachpb.Error
+	var err error
+	var positions []int
 
 	isReverse := ba.IsReverse()
 
-	// Truncate the request to range descriptor.
-	intersected, err := rs.Intersect(desc)
-	if err != nil {
-		return response{pErr: roachpb.NewError(err)}
-	}
-	truncBA, positions, err := truncate(ba, intersected)
-	if len(positions) == 0 && err == nil {
-		// This shouldn't happen in the wild, but some tests exercise it.
-		return response{
-			pErr: roachpb.NewErrorf("truncation resulted in empty batch on %s: %s", intersected, ba),
+	if needsTruncate {
+		// Truncate the request to range descriptor.
+		rs, err = rs.Intersect(desc)
+		if err != nil {
+			return response{pErr: roachpb.NewError(err)}
 		}
-	}
-	if err != nil {
-		return response{pErr: roachpb.NewError(err)}
+		ba, positions, err = truncate(ba, rs)
+		if len(positions) == 0 && err == nil {
+			// This shouldn't happen in the wild, but some tests exercise it.
+			return response{
+				pErr: roachpb.NewErrorf("truncation resulted in empty batch on %s: %s", rs, ba),
+			}
+		}
+		if err != nil {
+			return response{pErr: roachpb.NewError(err)}
+		}
 	}
 
 	// Start a retry loop for sending the batch to the range.
@@ -925,9 +937,9 @@ func (ds *DistSender) sendPartialBatch(
 		if desc == nil {
 			var descKey roachpb.RKey
 			if isReverse {
-				descKey = intersected.EndKey
+				descKey = rs.EndKey
 			} else {
-				descKey = intersected.Key
+				descKey = rs.Key
 			}
 			desc, evictToken, err = ds.getDescriptor(ctx, descKey, nil, isReverse)
 			if err != nil {
@@ -936,7 +948,7 @@ func (ds *DistSender) sendPartialBatch(
 			}
 		}
 
-		reply, pErr = ds.sendSingleRange(ctx, truncBA, desc)
+		reply, pErr = ds.sendSingleRange(ctx, ba, desc)
 
 		// If sending succeeded, return immediately.
 		if pErr == nil {
@@ -996,7 +1008,7 @@ func (ds *DistSender) sendPartialBatch(
 			// batch here would give a potentially larger response slice
 			// with unknown mapping to our truncated reply).
 			log.VEventf(ctx, 1, "likely split; resending batch to span: %s", tErr)
-			reply, pErr = ds.divideAndSendBatchToRanges(ctx, truncBA, intersected, batchIdx)
+			reply, pErr = ds.divideAndSendBatchToRanges(ctx, ba, rs, batchIdx)
 			return response{reply: reply, positions: positions, pErr: pErr}
 		}
 		break
@@ -1115,12 +1127,6 @@ func (ds *DistSender) sendToReplicas(
 	args roachpb.BatchRequest,
 	rpcContext *rpc.Context,
 ) (*roachpb.BatchResponse, error) {
-	if len(replicas) < 1 {
-		return nil, roachpb.NewSendError(
-			fmt.Sprintf("insufficient replicas (%d) to satisfy send request of %d",
-				len(replicas), 1))
-	}
-
 	var ambiguousError error
 	var haveCommit bool
 	// We only check for committed txns, not aborts because aborts may