- Fast path at start of `divideAndSendBatchToRanges` for case where
  all the key span is contained within a single range.
- Avoid batch truncation in fast path in `sendPartialBatch`.
- Fast path in `grpcTransport` for case of a single replica - send
  syncrhonously instead of via goroutine.
- Remove `DistSender.sendRPC`'s `context.WithCancel` as cancellation should
  properly be done in the transport, which is the component which understands
  if cancellation is necessary. gRPCTransport already has a cancellation
  mechanism, so we were doing this twice unnecessarily. In the case of the
  dist sender fast path (added to this commit) which prevents launching a
  goroutine in the event of a single replica, this means the transport can
  avoid starting any cancelable context.
- Remove unnecessary transaction clone from grpcTransport for local
  server case because after code audit, determined that there are no
  cases where `ba.Txn` is modified at `Store.Send` and below without
  first doing a similar shallow copy of `ba.Txn`.

I removed the shallow clone of the batch's transaction.  This is no
longer needed, based on a code audit. There is nowhere in the Store ->
Replica -> Raft evaluation path where we modify the batch's Txn
without first making a shallow copy of it.