@nvanbenschoten 2017-10-13
Reviewed 6 of 6 files at r1.
Review status: all files reviewed at latest revision, 5 unresolved discussions, all commit checks successful.

---

*[pkg/kv/dist_sender.go, line 900 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/19200#-KwI5_d8RJ4ilOMnMGYv:-KwI5_d8RJ4ilOMnMGYw:b86qfq4) ([raw file](https://github.com/cockroachdb/cockroach/blob/e797510985ac138e4972a51a0e67e00d6cf23e7d/pkg/kv/dist_sender.go#L900)):*
> ```Go
> 	evictToken *EvictionToken,
> 	batchIdx int,
> 	needsTruncate bool,
> ```

Could you add a comment describing what `needsTruncate` means?

---

*[pkg/kv/dist_sender.go, line 918 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/19200#-KwDBmw1v8Os9G3prIju-r1-918:-KwI67m1h_I23NXFybSf:bthczj4) ([raw file](https://github.com/cockroachdb/cockroach/blob/e797510985ac138e4972a51a0e67e00d6cf23e7d/pkg/kv/dist_sender.go#L918)):*
<details><summary><i>Previously, tamird (Tamir Duberstein) wrote…</i></summary><blockquote>

can this overwrite `rs` instead of assigning to `intersected`? seems like it'd reduce the possibility of doing the wrong thing. `ba` and `truncBA` seem to demonstrate the same oddity.
</blockquote></details>

Yeah, I think getting rid of `intersected` and `truncBA` would be cleaner. Just reassign to the previous values `if needsTruncate`.

---

*[pkg/kv/transport.go, line 192 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/19200#-KwI73K5dMm6BkgbT4OL:-KwI73K5dMm6BkgbT4OM:b194gmc) ([raw file](https://github.com/cockroachdb/cockroach/blob/e797510985ac138e4972a51a0e67e00d6cf23e7d/pkg/kv/transport.go#L192)):*
> ```Go
> 	// on context or launch in a goroutine.
> 	if len(gt.orderedClients) == 1 {
> 		done <- gt.send(ctx, client)
> ```

We no longer have the `NotLeaseHolderErrors` check here. Is it not possible for a single replica to temporarily lose an expiration-based lease if it gets slowed down for long enough? In that case, wouldn't we want to retry?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19200)*
<!-- Sent from Reviewable.io -->