@petermattis 2017-10-13
Review status: all files reviewed at latest revision, 7 unresolved discussions, all commit checks successful.

---

*[pkg/kv/transport.go, line 217 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/19200#-KwKyLfkWE5hizslqcUm:-KwKyLfkWE5hizslqcUn:b-oyx877) ([raw file](https://github.com/cockroachdb/cockroach/blob/359532ca24405996b21d0a8ab6e6c5549dde60b6/pkg/kv/transport.go#L217)):*
> ```Go
> 					clonedTxn := origTxn.Clone()
> 					client.args.Txn = &clonedTxn
> 				}
> ```

This might no longer be necessary, but we should pay attention to test race failures to make sure. 

---

*[pkg/kv/transport.go, line 220 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/19200#-KwKxjoRBxPi-9BCH9NZ:-KwKxjoRBxPi-9BCH9N_:b-m8o8er) ([raw file](https://github.com/cockroachdb/cockroach/blob/e797510985ac138e4972a51a0e67e00d6cf23e7d/pkg/kv/transport.go#L220)):*
> ```Go
> 			}
> 		}
> 		gt.setState(client.args.Replica, false /* pending */, retryable)
> ```

Curious why this code wasn't moved to `grpcTransport.send`. Doing so would address @nvanbenschoten's comment above.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19200)*
<!-- Sent from Reviewable.io -->