@bdarnell 2017-10-16
This generally looks good, but optimization PRs should nearly always be accompanied by benchmark numbers. And separate commits would also be helpful (both for reviewing and for future bisecting).

---

Reviewed 6 of 6 files at r1.
Review status: all files reviewed at latest revision, 7 unresolved discussions, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19200#-:-KwY39taWvWa2HebmISy:b-bj5tpb)*
<!-- Sent from Reviewable.io -->