@spencerkimball 2017-10-23
Performance on single node 99% read-percent KV load generator improved with this PR by ~10%.

BEFORE
```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   60.0s        0         568667         9477.5      3.4      2.8     10.0     18.9     92.3

BenchmarkLoadgenKV/readPercent=99/splits=0/concurrency=32/duration=1m0s	  568667	    105513 ns/op

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   60.0s        0         563939         9398.6      3.4      2.8     10.0     19.9    184.5

BenchmarkLoadgenKV/readPercent=99/splits=0/concurrency=32/duration=1m0s	  563939	    106399 ns/op

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   60.0s        0         573619         9560.0      3.3      2.8     10.0     18.9    121.6

BenchmarkLoadgenKV/readPercent=99/splits=0/concurrency=32/duration=1m0s	  573620	    104602 ns/op
```

AFTER
```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   60.0s        0         634057        10567.3      3.0      2.5      8.9     16.8    113.2

BenchmarkLoadgenKV/readPercent=99/splits=0/concurrency=32/duration=1m0s	  634057	     94631 ns/op

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   60.0s        0         629878        10497.7      3.0      2.5      8.9     17.8    159.4

BenchmarkLoadgenKV/readPercent=99/splits=0/concurrency=32/duration=1m0s	  629879	     95258 ns/op

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   60.0s        0         616276        10270.6      3.1      2.6      8.9     17.8    113.2

BenchmarkLoadgenKV/readPercent=99/splits=0/concurrency=32/duration=1m0s	  616276	     97365 ns/op
```